refactor and ensure test coverage hits both
