+++ converter.go
@@ -1,0 +1,5 @@
+package textplain
+
+type Converter interface {
+	Convert(string, int) (string, error)
+}
