+++ setup_test.go
@@ -8,6 +8,7 @@ import (
 )
 
 func runTestCases(t *testing.T, testCases []testCase) {
+
 	for _, tc := range testCases {
 		t.Run(tc.name, func(tt *testing.T) {
 			runTestCase(tt, tc)
@@ -16,7 +17,18 @@ func runTestCases(t *testing.T, testCases []testCase) {
 }
 
 func runTestCase(t *testing.T, tc testCase) {
-	result, err := textplain.Convert(tc.body, textplain.DefaultLineLength)
-	assert.Nil(t, err)
-	assert.Equal(t, tc.expect, result)
+
+	converters := map[string]textplain.Converter{
+		"regexp": textplain.NewRegexpConverter(),
+		"tree":   textplain.NewTreeConverter(),
+	}
+
+	for name, converter := range converters {
+		t.Run(name, func(tt *testing.T) {
+			result, err := converter.Convert(tc.body, textplain.DefaultLineLength)
+			assert.Nil(t, err)
+			assert.Equal(t, tc.expect, result)
+		})
+	}
+
 }
